增加登录认证功能

+++ src/main/java/cn/zucc/edu/club/entity/StuClub.java
@@ -28,7 +28,7 @@ public class StuClub implements Serializable {
 
     private Integer clubId;
 
-    private Integer stuId;
+    private Long stuId;
 
     private String applyContent;
