+++ src/main/java/cn/zucc/edu/club/mapper/StuClubMapper.java
@@ -27,6 +27,6 @@ public interface StuClubMapper extends BaseMapper<StuClub> {
 
     List<StuClub> searchApplyStuClub(int clubId);
 
-    StuClub getOneStuInClub(int clubId, int stuId);
+    StuClub getOneStuInClub(int clubId, Long stuId);
 
 }
